pyproject.toml
[project]
name = "drivenow-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "boto3>=1.34.0",
    "lxml>=4.9.0",
    "pillow>=10.0.0",
    "playwright>=1.40.0",
    "psycopg2-binary>=2.9.9",
    "python-dateutil>=2.8.2",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "selenium>=4.15.0",
    "sqlalchemy>=2.0.0",
    "webdriver-manager>=4.0.0",
]
